hybrid/requirements.txt
antlr4-python3-runtime==4.8
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
executing==2.0.1
fairseq @ file:///Users/vanloc1808/GitHub/cheapfakes_detection_icmr2024/fairseq
fonttools==4.51.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
hydra-core==1.0.7
idna==3.6
ipdb==0.13.13
ipython==8.23.0
jedi==0.19.1
kiwisolver==1.4.5
lxml==5.2.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
numpy==1.23.5
omegaconf==2.0.6
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
portalocker==2.8.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
sacrebleu==2.4.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
tqdm==4.66.2
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
